Feat: implementing middleware for with JWTs
